// app/src/api/portfolio/PortfolioClient.ts
export interface Comment {
  id: string;
  update_id: string;
  author: string;
  body: string;
  created_at: number;
}

export interface Event_CommentPosted {
  id: string;
}

export interface Event_CompanyFollowed {
  id: string;
}

export interface Event_MetricLogged {
  id: string;
}

export interface Event_UpdatePosted {
  id: string;
}

export interface Metric {
  id: string;
  company_name: string;
  metric_name: string;
  value: string;
  author: string;
  timestamp_ms: number;
}

export interface PortfolioState {
  updates: Record<string, Update>;
  metrics: Record<string, Metric>;
  comments: Record<string, Comment>;
  subscriptions: Record<string, Subscription>;
}

export interface Subscription {
  id: string;
  subscriber: string;
  company_name: string;
}

export interface Update {
  id: string;
  company_name: string;
  author: string;
  body: string;
  created_at: number;
}





export type AbiEvent =
  | { name: "UpdatePosted"; payload: Event_UpdatePosted }
  | { name: "MetricLogged"; payload: Event_MetricLogged }
  | { name: "CommentPosted"; payload: Event_CommentPosted }
  | { name: "CompanyFollowed"; payload: Event_CompanyFollowed }
;


export class PortfolioClient {
  private _mero: MeroJs;
  private _contextId: string;
  private _executorPublicKey: string;

  constructor(mero: MeroJs, contextId: string, executorPublicKey: string) {
    this._mero = mero;
    this._contextId = contextId;
    this._executorPublicKey = executorPublicKey;
  }

  /**
   * init
   */
  public async init(): Promise<void> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'init', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as void;
  }

  /**
   * post_update
   */
  public async postUpdate(params: { company_name: string; body: string }): Promise<string> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'post_update', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as string;
  }

  /**
   * get_updates
   */
  public async getUpdates(): Promise<Update[]> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'get_updates', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as Update[];
  }

  /**
   * log_metric
   */
  public async logMetric(params: { company_name: string; metric_name: string; value: string }): Promise<string> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'log_metric', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as string;
  }

  /**
   * get_latest_metrics
   */
  public async getLatestMetrics(): Promise<Metric[]> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'get_latest_metrics', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as Metric[];
  }

  /**
   * post_comment
   */
  public async postComment(params: { update_id: string; body: string }): Promise<string> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'post_comment', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as string;
  }

  /**
   * get_comments
   */
  public async getComments(params: { update_id: string }): Promise<Comment[]> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'get_comments', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as Comment[];
  }

  /**
   * follow_company
   */
  public async followCompany(params: { company_name: string }): Promise<string> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'follow_company', argsJson: params, executorPublicKey: this._executorPublicKey });
    return response as string;
  }

  /**
   * get_subscriptions
   */
  public async getSubscriptions(): Promise<Subscription[]> {
    const response = await this._mero.rpc.execute({ contextId: this._contextId, method: 'get_subscriptions', argsJson: {}, executorPublicKey: this._executorPublicKey });
    return response as Subscription[];
  }

}